requirements.txt
mistralai==1.6.0
requests==2.32.3
python-dotenv==1.1.0